[feature] Add support for wallet-connect compatible wallets (#88)
* [feature] Add support for wallet-connect compatible wallets

* [refactor] Protect from failures when no accounts are found after wallet connection

* [bugfix] Backport web3 version to 1.2.x

There is an issue with web3 1.3.0 see here:
https://github.com/ethereum/web3.js/issues/3790

This prevents connections using wallet connect option

* [bugfix] Use returned Web3Modal provider

* [bugfix] Throttle calls to tokens balance/allowance

Since we manage a list of 1500+ tokens, the calls made in
parallel to get balance/allowance info were causing
resource issues on the infura endpoint.

Changed this to batch calls into buckets of 200. Each
bucket execution is separated by 10s. And the whole cycle
is re-run each minute

* [ux] refresh orderbook every 10s instead of every 1s

To reduce the amount of calls to 0x and coingecko APIs
and also refresh the UI less frequently for better usability

* [ux] do not cache provider to enable user switching between wallets

* [bugfix] Only fetch zeroXcontract address when wallet is connected

* [layout] Make modal display on top of all ui components