[ux] Improve order submit user feedback (#107)
* [ux] Improve order submit user feedback

* [bugfix] fix filling order message amount value

* [perf] Synchronize tokens balances/allowance only when needed

Currently the balances/allowances are synchronized in
frequent process each 1 minute. This causes lag issues
on interaction with wallets.

Restricting updates to only:
- after token list is fetched
- once at each wallet account change

* [ux] Add token loaders

As long as token balance/allowance is not updated
loader will replace balances in token selector
as well as the MAX button in the order form

* [ux] Default to Rpc provider for adding tokens by address

On first load using token pair address link, the tokens
may not be present in the tokens list. In that case
add the token to the list by querying its contract
using the default rpc provider targeting infura.

This is needed as a wallet connection may not always
be available

* [ux] Show CONNECT WALLET on form button when needed

* [ux] Check wallet is connected before set token as initialized

As wallet connection can be reset between balance update
and allowance call, we need to recheck that wallet connection
still exists just before setting a token as initialized

* [ux] Show token loader only when wallet is connected

When wallet not connected, display a connect wallet
message on the deactivated form button, and trigger
wallet connection if user clicks on MAX button

* [ux] Update token display after token refresh

Even if no balance/info found trigger token
display update to remove loader