Add support for market orders (#110)
* [feature] Add support for market orders (swaps)

- Included improvements in user feedback messages
- Add loader on token selector when searching
  for a token with an address
- Tweak logo and blocky
- Use blocky icon as token logo when logo not found
- Put an event system to manage interactions
  between different UI elements

* [bugfix] Always encore amount/fee in digits when calling 0x swap API

* [refactor] Use same method for updating allowance everywhere

* [bugfix] Always set takerAddress in swap quote call

This should lead to a better gas estimation in the
quote reply

* [bugfix] Correctly set all swap critical params

takerAddress was always empty causing the transaction
to fail in some cases. Also set splippage to appropriate
value as well as other params

* [layout] Remove blocky icon shadow

* [refactor] Use price api instead of quote for frequent market price queries

* [ux] Add missing translations for MARKET/LIMIT tabs

* [ux] Show 50 tokens by default

To reduce token list rendering impact on performance
reduce default token list display to 50. The search
will still apply to the whole token list, and user
can access all tokens present in the list

* [bugfix] Reset amount/price values when selecting a new token

* [feature] Add support for ETH in market orders

* [ux] Forbid selecting same token for sell & buy

* [ux] Reset price waiting for refresh after token change

* [ux] When selling ETH underestimate max selection with 0.1

This to account for transaction cost and avoid 0x API
rejection when requesting quote

* [ux] Refresh swap price when user changes amounts

Use the actual amounts to request the 0x quote, this
way it displays an accurate price, as close as possible
to the actual one at submit time

* [bugfix] Correct fee encoding on swap quote query

* [ux] Refresh price on buy amount changes when in market mode

* [feature] Add Basis.Gold tokens to default token list

* [ux] Add token logo for Basis Gold

* [ux] Reset values & update order form in parallel

* [feature] Add support for ETH WRAP/UNWRAP with 0 fees